Add functionality for setup button
When the setup button gets clicked, the SetupWindow frame will be
added to the Panedwindow controller.

For now( since SetupWindow is the only frame to add/remove from
the controller) when the button is pressed, the function will check
if there is only one pane active, then add the SetupWindow. If it is
not only one pane, then it will close the last pane from the controller.