Class grid vars, Lock/Unlock functionality, & data population
Add class variable to hold the grid data for every input.
This was done to have an easier way to add and remove widgets from the frame.

Create function to make the Lock/Unlock button work.

Create function to populate setup data if it was already created and saved to the database previously.