Add way to check if App and DB are out of date.
The program now checks if the app is out of date checking a gist and comparing it to the code in __init__.py.
If it is out of date, then a window pops up saying to go to the releases and update the app.
If it is not out of date then it continues to the database check.

The database check also uses the gist to check what the newest version of the DB is.
If the app has an old DB version, then the new sql files are downloaded from the gist and the DB is updated.

#12