Create a way to cancel an upload.
After an upload is started, the Start Upload button gets replaced with a Cancel Upload button.
When the Cancel Upload button is pressed, the program waits for a new itteration through the files and runs a function that resets the progressbars, changes the lavel text to Upload Canceled, enables the widgets, and re-adds the Start Upload button.

#8